requirements.txt
pandas
numpy
sqlalchemy
pymysql
python-dotenv
requests
scikit-learn
pyspark
matplotlib
seaborn
plotly